login: Rewrite utmp/wtmp locking mechanism (BZ #24492)
The utmp locking mechanism locks the utmp database file (either the
default UTMP_FILE/WTMP_FILE or the one passed on utmpname).  For default
UTMP_FILE/WTMP_FILE, the file is usually readable by non-root users and
thus it can be locked as well.

The current implementation mitigates by using an alarm with a non
configurable TIMEOUT to force the lock to fail if the timer expiries.
Although it makes the putut{x}line/updwtmp{x} to not block indefinitely,
it allows an unprivileged process to deny updates on the databases.

This patch fixes it by locking a different file and only on symbols that
actually update the databases (putut{x}line/updwtmp{x}).  The lock file
path is constructed based on the selected database:

  - for default ones (UTMP_FILE/WTMP_FILE) the UTMP_FILE ".lock" and
    WTMP_FILE ".lock" file is used.

  - Otherwise the lock file is constructed by appending the '.lock'
    suffix on the file name set by utmpname.

Also, the lock is not create for default databases since the '/var/lock'
usually has permissive access and any process could create them.
Rather, it is expected the lock files to be created with more
restrictive permission so unprivileged processes can't lock it.

The reads are done without locking, which theoretically means that a
read could see a partially-updated record.  I don't see a easy way to
avoid it while allowing the utmp databases to be accessible to all users
and without making the utmpx access API to use a broker to access
the database.  In any case, I think this issue is less of an issue than
missing updates from priviledege processes.

This patch is based on top of the utmp/utmp 64-bit support [1]

Checked on x86_64-linux-gnu and i686-linux-gnu.

[1] https://sourceware.org/pipermail/libc-alpha/2020-August/116850.html